+++ src/interpreter/gc/unique_id.rs
@@ -2,13 +2,14 @@ use std::{
     fmt::{Display, Formatter},
     sync::atomic::{AtomicUsize, Ordering},
 };
+use std::sync::atomic::AtomicU32;
 
 use serde::{Deserialize, Serialize};
 
 /// A unique ID, suitable for uniquely identifying an object that contains
 /// references that could be garbage-collected.
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
-pub struct UniqueId(usize);
+pub struct UniqueId(u32);
 
 impl UniqueId {
     /// Create a new [`UniqueId`].
@@ -29,8 +30,8 @@ impl Display for UniqueId {
     }
 }
 
-impl AsRef<usize> for UniqueId {
-    fn as_ref(&self) -> &usize {
+impl AsRef<u32> for UniqueId {
+    fn as_ref(&self) -> &u32 {
         &self.0
     }
 }
@@ -39,6 +40,6 @@ impl AsRef<usize> for UniqueId {
 /// This ID is used so we can avoid track objects that have been marked,
 /// to avoid infinitely processing a looping graph.
 fn next_id() -> UniqueId {
-    static COUNTER: AtomicUsize = AtomicUsize::new(1);
+    static COUNTER: AtomicU32 = AtomicU32::new(1);
     UniqueId(COUNTER.fetch_add(1, Ordering::Relaxed))
 }
