+++ src/interpreter/object_space.rs
@@ -313,7 +313,7 @@ mod tests {
         let mut processed = BitSet::new();
         space.mark(&mut marked, &mut processed).unwrap();
 
-        assert!(processed.contains(*array_id.as_ref()));
+        assert!(processed.contains(*array_id.as_ref() as usize));
     }
 
     #[test]
